Clean up the pmap code a little
Add a new function to check if the entry is dirty. We don't yet track
this correctly so it will return true for all writable pages.

Check if the access flag is set to determine if a page has been accessed.
This will also need to be tracked in software as we are expected to clear
this bit when we create the map and set it in an exception handler when
the access happens.