Add support for handling the VFP context
We restore the context in a userland VFP exception, i.e. when userland
attempts to access the VFP registers or floating-point special registers.
At the same time we keep a record of the last CPU we ran the thread ran
the VFP context on and the last VFP context we ran on the current cpu.
If these line up we can skip restoring the context as it will sill be
valid.

We store the context on context switches. To be lazy we check if the
VFP is enabled before doing this, and if not can skip it as this means
the thread hasn't used the unit. If it had we would have enabled the
VFP for the access.

We also store the VFP context around signals. To store it we need to
make sure the current thread has a valid context, then copy it to
the mcontext_t. On the return from a signal handler we then restore
the old context from the mcontext_t.